+++ packages/plugin-metro/src/lib/getReactNativeDeps.ts
